requirements.txt
APScheduler==3.11.1
blinker==1.9.0
click==8.3.0
Flask==3.1.2
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pyotp==2.9.0
python-dotenv==1.2.1
SQLAlchemy==2.0.44
typing_extensions==4.15.0
tzlocal==5.3.1
Werkzeug==3.1.3
FlaskFlask-SQLAlchemyFlask-LoginFlask-MailAPSchedulerpython-dotenvpyotpFlask-WTFemail_validator
